requirements.txt
aioboto3==15.2.0
aiobotocore==2.24.2
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aioitertools==0.12.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.30.0
attrs==25.4.0
bcrypt==3.2.2
boto3==1.40.18
botocore==1.40.18
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.118.3
fastapi-cli==0.0.13
fastapi-cloud-cli==0.3.1
frozenlist==1.8.0
geoip2==5.1.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.10
IP2Location==8.11.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
maxminddb==2.8.2
mdurl==0.1.2
multidict==6.7.0
orjson==3.11.3
passlib==1.7.4
pillow==11.3.0
propcache==0.4.1
psutil==7.1.0
psycopg==3.2.11
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.0
pydantic-extra-types==2.10.6
pydantic-settings==2.11.0
pydantic_core==2.41.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.3
qrcode==8.2
redis==6.4.0
requests==2.32.5
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.0
rsa==4.9.1
s3transfer==0.13.1
segno==1.6.6
sentry-sdk==2.41.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
ujson==5.11.0
urllib3==2.5.0
user-agents==2.2.0
uvicorn==0.37.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
wrapt==1.17.3
yarl==1.22.0
